Khush Patel87

+++ crm/crm_clients.php
@@ -1,4 +1,6 @@
-<?php include 'crm_header.php'; ?>
+<?php include 'crm_header.php';
+include 'db_connection.php'; ?>
+
 <div class="container mt-4">
     <h2>Client Management</h2>
     <a href="crm_add_client.php" class="btn btn-primary mb-3">Add Client</a>
@@ -26,7 +28,7 @@
                     <td>{$row['company']}</td>
                     <td>
                         <a href='crm_edit_client.php?id={$row['id']}' class='btn btn-warning btn-sm'>Edit</a>
-                        <a href='crm_delete_client.php?id={$row['id']}' class='btn btn-danger btn-sm'>Delete</a>
+                        <a href='crm_delete_client.php?id={$row['id']}' class='btn btn-danger btn-sm' onclick='return confirm(\"Are you sure?\")'>Delete</a>
                     </td>
                 </tr>";
             }
@@ -34,4 +36,5 @@
         </tbody>
     </table>
 </div>
+
 <?php include 'crm_footer.php'; ?>
